file uploader modification for react-refresh rule
