+++ frontend/lib/src/components/widgets/FileUploader/utils.test.ts
@@ -1,0 +1,34 @@
+/**
+ * Copyright (c) Streamlit Inc. (2018-2022) Snowflake Inc. (2022-2025)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { getAccept, STREAMLIT_MIME_TYPE } from "./utils"
+
+describe("FileUploader utils", () => {
+  describe("getAccept", () => {
+    it("returns undefined when no extensions provided", () => {
+      expect(getAccept([])).toBeUndefined()
+    })
+
+    it("returns accept object with MIME type when extensions provided", () => {
+      const extensions = [".jpg", ".png"]
+      const result = getAccept(extensions)
+
+      expect(result).toEqual({
+        [STREAMLIT_MIME_TYPE]: extensions,
+      })
+    })
+  })
+})
